third_party/WebKit/LayoutTests/fast/workers/worker-formdata-expected.txt

[Worker] Test FormData interface object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: ./resources/worker-formdata.js
PASS [Worker] FormData is defined.
PASS [Worker] FormData.length is 0
PASS [Worker] formData is non-null.
PASS [Worker] FormData.prototype.hasOwnProperty('append') is true
PASS [Worker] formData.append('key', 'value'); did not throw exception.
PASS [Worker] blob is non-null.
PASS [Worker] formData.append('key', blob); did not throw exception.
PASS [Worker] formData.append('key', blob, 'filename'); did not throw exception.
PASS [Worker] postMessage(formData) threw exception DataCloneError: Failed to execute 'postMessage' on 'DedicatedWorkerGlobalScope': FormData object could not be cloned..
PASS successfullyParsed is true

TEST COMPLETE
